+++ common.py
@@ -1,0 +1,104 @@
+def kill_process(process_name):
+    for proc in psutil.process_iter():
+        if proc.name == process_name:
+            proc.kill()
+
+def get_ip_address(ifname):
+    try:
+        s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+        return socket.inet_ntoa(fcntl.ioctl(s.fileno(),
+            0x8915,  # SIOCGIFADDR
+            struct.pack('256s', ifname[:15])
+        )[20:24])
+    except Exception:
+        print "Cannot get IP Address for Interface %s" % ifname
+        sys.exit(1)
+
+def delete_file(file_path):
+    if os.path.isfile(file_path):
+        os.remove(file_path)
+    else:
+        print("Error: %s file not found" % file_path)
+
+def write_to_file(file_path, content):
+    open(file_path, "a").write(content)
+
+def add_to_conf(conf_file, section, param, val):
+    config = iniparse.ConfigParser()
+    config.readfp(open(conf_file))
+    if not config.has_section(section):
+        config.add_section(section)
+        val += '\n'
+    config.set(section, param, val)
+    with open(conf_file, 'w') as f:
+        config.write(f)
+
+def delete_from_conf(conf_file, section, param):
+    config = iniparse.ConfigParser()
+    config.readfp(open(conf_file))
+    if param is None:
+        config.remove_section(section)
+    else:
+        config.remove_option(section, param)
+    with open(conf_file, 'w') as f:
+        config.write(f)
+
+def get_from_conf(conf_file, section, param):
+    config = iniparse.ConfigParser()
+    config.readfp(open(conf_file))
+    if param is None:
+        raise Exception("parameter missing")
+    else:
+        return config.get(section, param)
+
+def print_format(string):
+    print "+%s+" %("-" * len(string))
+    print "|%s|" % string
+    print "+%s+" %("-" * len(string))
+
+def execute(command, display=False):
+    print_format("Executing  :  %s " % command)
+    process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    if display:
+        while True:
+            nextline = process.stdout.readline()
+            if nextline == '' and process.poll() != None:
+                break
+            sys.stdout.write(nextline)
+            sys.stdout.flush()
+
+        output, stderr = process.communicate()
+        exitCode = process.returncode
+    else:
+        output, stderr = process.communicate()
+        exitCode = process.returncode
+
+    if (exitCode == 0):
+        return output.strip()
+    else:
+        print "Error", stderr
+        print "Failed to execute command %s" % command
+        print exitCode, output
+        raise Exception(output)
+
+def initialize_system():
+    if not os.geteuid() == 0:
+        sys.exit('Please re-run the script with root user')
+
+    execute("apt-get clean" , True)
+    execute("apt-get autoclean -y" , True)
+    execute("apt-get update -y" , True)
+    execute("apt-get install ubuntu-cloud-keyring python-setuptools python-iniparse python-psutil -y", True)
+    delete_file("/etc/apt/sources.list.d/icehouse.list")
+    execute("echo deb http://ubuntu-cloud.archive.canonical.com/ubuntu precise-updates/icehouse main >> /etc/apt/sources.list.d/icehouse.list")
+    execute("apt-get update -y", True)
+    execute("apt-get install vlan bridge-utils -y", True)
+    execute("sed -i 's/#net.ipv4.ip_forward=1/net.ipv4.ip_forward=1/' /etc/sysctl.conf")
+
+    global iniparse
+    if iniparse is None:
+        iniparse = __import__('iniparse')
+
+    global psutil
+    if psutil is None:
+        psutil = __import__('psutil')
