+++ icehouse-computenode.py
@@ -6,121 +6,13 @@ import fcntl
 import struct
 import socket
 import subprocess
+from common import *
 
 # These are module names which are not installed by default.
 # These modules will be loaded later after downloading
 iniparse = None
 psutil = None
 
-
-def kill_process(process_name):
-    for proc in psutil.process_iter():
-        if proc.name == process_name:
-            proc.kill()
-
-
-def delete_file(file_path):
-    if os.path.isfile(file_path):
-        os.remove(file_path)
-    else:
-        print("Error: %s file not found" % file_path)
-
-def write_to_file(file_path, content):
-    open(file_path, "a").write(content)
-
-def add_to_conf(conf_file, section, param, val):
-    config = iniparse.ConfigParser()
-    config.readfp(open(conf_file))
-    if not config.has_section(section):
-        config.add_section(section)
-        val += '\n'
-    config.set(section, param, val)
-    with open(conf_file, 'w') as f:
-        config.write(f)
-
-
-def delete_from_conf(conf_file, section, param):
-    config = iniparse.ConfigParser()
-    config.readfp(open(conf_file))
-    if param is None:
-        config.remove_section(section)
-    else:
-        config.remove_option(section, param)
-    with open(conf_file, 'w') as f:
-        config.write(f)
-
-
-def get_from_conf(conf_file, section, param):
-    config = iniparse.ConfigParser()
-    config.readfp(open(conf_file))
-    if param is None:
-        raise Exception("parameter missing")
-    else:
-        return config.get(section, param)
-
-def print_format(string):
-    print "+%s+" %("-" * len(string))
-    print "|%s|" % string
-    print "+%s+" %("-" * len(string))
-
-def execute(command, display=False):
-    print_format("Executing  :  %s " % command)
-    process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    if display:
-        while True:
-            nextline = process.stdout.readline()
-            if nextline == '' and process.poll() != None:
-                break
-            sys.stdout.write(nextline)
-            sys.stdout.flush()
-
-        output, stderr = process.communicate()
-        exitCode = process.returncode
-    else:
-        output, stderr = process.communicate()
-        exitCode = process.returncode
-
-    if (exitCode == 0):
-        return output.strip()
-    else:
-        print "Error", stderr
-        print "Failed to execute command %s" % command
-        print exitCode, output
-        raise Exception(output)
-
-def get_ip_address(ifname):
-        try:
-            s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            return socket.inet_ntoa(fcntl.ioctl(s.fileno(),
-                0x8915,  # SIOCGIFADDR
-                struct.pack('256s', ifname[:15])
-            )[20:24])
-        except Exception:
-            print "Cannot get IP Address for Interface %s" % ifname
-            sys.exit(1)
-
-
-def initialize_system():
-    if not os.geteuid() == 0:
-        sys.exit('Please re-run the script with root user')
-
-    execute("apt-get clean" , True)
-    execute("apt-get autoclean -y" , True)
-    execute("apt-get update -y" , True)
-    execute("apt-get install ubuntu-cloud-keyring python-setuptools python-iniparse python-psutil -y", True)
-    delete_file("/etc/apt/sources.list.d/icehouse.list")
-    execute("echo deb http://ubuntu-cloud.archive.canonical.com/ubuntu precise-updates/icehouse main >> /etc/apt/sources.list.d/icehouse.list")
-    execute("apt-get update -y", True)
-    execute("apt-get install vlan bridge-utils -y", True)	  
-    execute("sed -i 's/#net.ipv4.ip_forward=1/net.ipv4.ip_forward=1/' /etc/sysctl.conf")
-
-    global iniparse
-    if iniparse is None:
-        iniparse = __import__('iniparse')
-
-    global psutil
-    if psutil is None:
-        psutil = __import__('psutil')
 #=================================================================================
 #==================   Components Installation Starts Here ========================
 #=================================================================================
@@ -146,14 +38,12 @@ def install_and_configure_nova():
     execute("apt-get install qemu-kvm libvirt-bin python-libvirt -y")
     execute("apt-get install nova-compute-kvm novnc -y", True)
 
-
     add_to_conf(nova_paste_conf, "filter:authtoken", "auth_host", ip_address_mgnt)
     add_to_conf(nova_paste_conf, "filter:authtoken", "auth_port", "35357")
     add_to_conf(nova_paste_conf, "filter:authtoken", "auth_protocol", "http")
     add_to_conf(nova_paste_conf, "filter:authtoken", "admin_tenant_name", "service")
     add_to_conf(nova_paste_conf, "filter:authtoken", "admin_user", "nova")
     add_to_conf(nova_paste_conf, "filter:authtoken", "admin_password", "nova")
-
 
     add_to_conf(nova_conf, "DEFAULT", "logdir", "/var/log/nova")
     add_to_conf(nova_conf, "DEFAULT", "verbose", "True")
@@ -190,8 +80,6 @@ def install_and_configure_nova():
     execute("service libvirt-bin restart", True)
     execute("service nova-compute restart", True)
    
-
-
 def install_and_configure_ovs():
     neutron_conf = "/etc/neutron/neutron.conf"
     neutron_paste_conf = "/etc/neutron/api-paste.ini"
@@ -199,8 +87,6 @@ def install_and_configure_ovs():
     execute("apt-get install openvswitch-switch openvswitch-datapath-dkms -y", True)
 	
     execute("ovs-vsctl --may-exist add-br br-int")
-    execute("ovs-vsctl --may-exist add-br br-eth2") 
-    execute("ovs-vsctl --may-exist add-port br-eth2 eth2")
   
     execute("apt-get install neutron-plugin-openvswitch-agent -y", True)
 
@@ -220,20 +106,16 @@ def install_and_configure_ovs():
     add_to_conf(neutron_paste_conf, "filter:authtoken", "admin_user", "neutron")
     add_to_conf(neutron_paste_conf, "filter:authtoken", "admin_password", "neutron")
     
-    add_to_conf(neutron_plugin_conf, "securitygroup", "firewall_driver", "neutron.agent.linux.iptables_firewall.OVSHybridIptablesFirewallDriver")
     add_to_conf(neutron_plugin_conf, "DATABASE", "sql_connection", "mysql://neutron:neutron@%s/neutron"%ip_address_mgnt)
-    add_to_conf(neutron_plugin_conf, "OVS", "bridge_mappings", "physnet1:br-eth2")
-    add_to_conf(neutron_plugin_conf, "OVS", "tenant_network_type", "vlan")
-    add_to_conf(neutron_plugin_conf, "OVS", "network_vlan_ranges", "physnet1:1000:2999")
+    add_to_conf(neutron_plugin_conf, "OVS", "enable_tunneling", "True")
+    add_to_conf(neutron_plugin_conf, "OVS", "local_ip", my_ip)
+    add_to_conf(neutron_plugin_conf, "OVS", "tunnel_type", "gre")
     add_to_conf(neutron_plugin_conf, "OVS", "integration_bridge", "br-int")
+    add_to_conf(neutron_plugin_conf, "securitygroup", "firewall_driver", "neutron.agent.linux.iptables_firewall.OVSHybridIptablesFirewallDriver")
 
     execute("service neutron-plugin-openvswitch-agent restart", True)
-   
 
 initialize_system()
 install_and_configure_ntp()
 install_and_configure_nova()
 install_and_configure_ovs()
-
-
-
