+++ icehouse-ctrl-ml2.py
@@ -6,6 +6,8 @@ import fcntl
 import struct
 import socket
 import subprocess
+import common
+from common import *
 
 # These are module names which are not installed by default.
 # These modules will be loaded later after downloading
@@ -15,117 +17,12 @@ psutil = None
 mysql_password = "secret"
 service_tenant = None
 
-def kill_process(process_name):
-    for proc in psutil.process_iter():
-        if proc.name == process_name:
-            proc.kill()
-
-def get_ip_address(ifname):
-        try:
-            s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            return socket.inet_ntoa(fcntl.ioctl(s.fileno(),
-                0x8915,  # SIOCGIFADDR
-                struct.pack('256s', ifname[:15])
-            )[20:24])
-        except Exception:
-            print "Cannot get IP Address for Interface %s" % ifname
-            sys.exit(1)
-
-def delete_file(file_path):
-    if os.path.isfile(file_path):
-        os.remove(file_path)
-    else:
-        print("Error: %s file not found" % file_path)
-
-def write_to_file(file_path, content):
-    open(file_path, "a").write(content)
-
-def add_to_conf(conf_file, section, param, val):
-    config = iniparse.ConfigParser()
-    config.readfp(open(conf_file))
-    if not config.has_section(section):
-        config.add_section(section)
-        val += '\n'
-    config.set(section, param, val)
-    with open(conf_file, 'w') as f:
-        config.write(f)
-
-
-def delete_from_conf(conf_file, section, param):
-    config = iniparse.ConfigParser()
-    config.readfp(open(conf_file))
-    if param is None:
-        config.remove_section(section)
-    else:
-        config.remove_option(section, param)
-    with open(conf_file, 'w') as f:
-        config.write(f)
-
-
-def get_from_conf(conf_file, section, param):
-    config = iniparse.ConfigParser()
-    config.readfp(open(conf_file))
-    if param is None:
-        raise Exception("parameter missing")
-    else:
-        return config.get(section, param)
-
-def print_format(string):
-    print "+%s+" %("-" * len(string))
-    print "|%s|" % string
-    print "+%s+" %("-" * len(string))
-
-def execute(command, display=False):
-    print_format("Executing  :  %s " % command)
-    process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    if display:
-        while True:
-            nextline = process.stdout.readline()
-            if nextline == '' and process.poll() != None:
-                break
-            sys.stdout.write(nextline)
-            sys.stdout.flush()
-
-        output, stderr = process.communicate()
-        exitCode = process.returncode
-    else:
-        output, stderr = process.communicate()
-        exitCode = process.returncode
-
-    if (exitCode == 0):
-        return output.strip()
-    else:
-        print "Error", stderr
-        print "Failed to execute command %s" % command
-        print exitCode, output
-        raise Exception(output)
-
 
 def execute_db_commnads(command):
     cmd = """mysql -uroot -p%s -e "%s" """ % (mysql_password, command)
     output = execute(cmd)
     return output
 
-
-def initialize_system():
-    if not os.geteuid() == 0:
-        sys.exit('Please re-run the script with root user')
-
-    execute("apt-get clean" , True)
-    execute("apt-get autoclean -y" , True)
-    execute("apt-get update -y" , True)
-    execute("apt-get install ubuntu-cloud-keyring python-setuptools python-iniparse python-psutil -y", True)
-    delete_file("/etc/apt/sources.list.d/icehouse.list")
-    execute("echo deb http://ubuntu-cloud.archive.canonical.com/ubuntu precise-updates/icehouse main >> /etc/apt/sources.list.d/icehouse.list")
-    execute("apt-get update -y", True)
-
-    global iniparse
-    if iniparse is None:
-        iniparse = __import__('iniparse')
-
-    global psutil
-    if psutil is None:
-        psutil = __import__('psutil')
 #=================================================================================
 #==================   Components Installation Starts Here ========================
 #=================================================================================
@@ -138,7 +35,6 @@ def install_rabbitmq():
     execute("service rabbitmq-server restart", True)
     time.sleep(2)
 
-
 def install_database():
     os.environ['DEBIAN_FRONTEND'] = 'noninteractive'
     execute("apt-get install mysql-server python-mysqldb mysql-client-5.5 -y", True)
@@ -149,8 +45,6 @@ def install_database():
         execute("mysqladmin -u root password %s" % mysql_password)
     except Exception:
         print " Mysql Password already set as : %s " % mysql_password
-
-
 
 def _create_keystone_users():
     os.environ['SERVICE_TOKEN'] = 'ADMINTOKEN'
@@ -164,14 +58,11 @@ def _create_keystone_users():
     admin_role = execute("keystone role-create --name admin|grep ' id '|awk '{print $4}'")
     execute("keystone user-role-add --user_id %s --tenant_id %s --role_id %s" % (admin_user, admin_tenant, admin_role))
 
-
     service_tenant = execute("keystone tenant-create --name service --description 'Service Tenant' --enabled true |grep ' id '|awk '{print $4}'")
-
 
     #keystone
     keystone_service = execute("keystone service-create --name=keystone --type=identity --description='Keystone Identity Service'|grep ' id '|awk '{print $4}'")
     execute("keystone endpoint-create --region region --service_id=%s --publicurl=http://%s:5000/v2.0 --internalurl=http://%s:5000/v2.0 --adminurl=http://%s:35357/v2.0" % (keystone_service, ip_address,ip_address_mgnt,ip_address_mgnt))
-
 
     #Glance
     glance_user = execute("keystone user-create --tenant_id %s --name glance --pass glance --enabled true|grep ' id '|awk '{print $4}'" % service_tenant)
@@ -180,14 +71,12 @@ def _create_keystone_users():
     glance_service = execute("keystone service-create --name=glance --type=image --description='Glance Image Service'|grep ' id '|awk '{print $4}'")
     execute("keystone endpoint-create --region region --service_id=%s --publicurl=http://%s:9292/v2 --internalurl=http://%s:9292/v2 --adminurl=http://%s:9292/v2" % (glance_service, ip_address,ip_address_mgnt,ip_address_mgnt))
 
-
     #nova
     nova_user = execute("keystone user-create --tenant_id %s --name nova --pass nova --enabled true|grep ' id '|awk '{print $4}'" % service_tenant)
     execute("keystone user-role-add --user_id %s --tenant_id %s --role_id %s" % (nova_user, service_tenant, admin_role))
 
     nova_service = execute("keystone service-create --name=nova --type=compute --description='Nova Compute Service'|grep ' id '|awk '{print $4}'")
     execute("keystone endpoint-create --region region --service_id=%s --publicurl='http://%s:8774/v2/$(tenant_id)s' --internalurl='http://%s:8774/v2/$(tenant_id)s' --adminurl='http://%s:8774/v2/$(tenant_id)s'" % (nova_service, ip_address,ip_address_mgnt,ip_address_mgnt))
-
 
     #neutron
     neutron_user = execute("keystone user-create --tenant_id %s --name neutron --pass neutron --enabled true|grep ' id '|awk '{print $4}'" % service_tenant)
@@ -205,7 +94,6 @@ def _create_keystone_users():
     write_to_file(adminrc, "export OS_AUTH_URL=http://%s:5000/v2.0\n" %ip_address)
 
 
-
 def install_and_configure_keystone():
     keystone_conf = "/etc/keystone/keystone.conf"
 
@@ -215,7 +103,6 @@ def install_and_configure_keystone():
     execute_db_commnads("GRANT ALL PRIVILEGES ON keystone.* TO 'keystone'@'localhost' IDENTIFIED BY 'keystone';")
 
     execute("apt-get install keystone -y", True)
-
 
     add_to_conf(keystone_conf, "DEFAULT", "admin_token", "ADMINTOKEN")
     add_to_conf(keystone_conf, "DEFAULT", "admin_port", 35357)
@@ -228,8 +115,6 @@ def install_and_configure_keystone():
 
     time.sleep(3)
     _create_keystone_users()
-
-
 
 def install_and_configure_glance():
     glance_api_conf = "/etc/glance/glance-api.conf"
@@ -244,7 +129,6 @@ def install_and_configure_glance():
 
     execute("apt-get install glance -y", True)
 
-
     add_to_conf(glance_api_paste_conf, "filter:authtoken", "auth_host", ip_address)
     add_to_conf(glance_api_paste_conf, "filter:authtoken", "auth_port", "35357")
     add_to_conf(glance_api_paste_conf, "filter:authtoken", "auth_protocol", "http")
@@ -252,14 +136,12 @@ def install_and_configure_glance():
     add_to_conf(glance_api_paste_conf, "filter:authtoken", "admin_user", "glance")
     add_to_conf(glance_api_paste_conf, "filter:authtoken", "admin_password", "glance")
 
-
     add_to_conf(glance_registry_paste_conf, "filter:authtoken", "auth_host", ip_address)
     add_to_conf(glance_registry_paste_conf, "filter:authtoken", "auth_port", "35357")
     add_to_conf(glance_registry_paste_conf, "filter:authtoken", "auth_protocol", "http")
     add_to_conf(glance_registry_paste_conf, "filter:authtoken", "admin_tenant_name", "service")
     add_to_conf(glance_registry_paste_conf, "filter:authtoken", "admin_user", "glance")
     add_to_conf(glance_registry_paste_conf, "filter:authtoken", "admin_password", "glance")
-
 
     add_to_conf(glance_api_conf, "DEFAULT", "sql_connection", "mysql://glance:glance@localhost/glance")
     add_to_conf(glance_api_conf, "paste_deploy", "flavor", "keystone")
@@ -276,10 +158,6 @@ def install_and_configure_glance():
 
     execute("service glance-api restart", True)
     execute("service glance-registry restart", True)
-	
-
-
-
 
 def install_and_configure_nova():
     nova_conf = "/etc/nova/nova.conf"
@@ -292,14 +170,12 @@ def install_and_configure_nova():
 
     execute("apt-get install nova-api nova-cert nova-scheduler nova-conductor novnc nova-consoleauth nova-novncproxy -y", True)
 
-
     add_to_conf(nova_paste_conf, "filter:authtoken", "auth_host", ip_address)
     add_to_conf(nova_paste_conf, "filter:authtoken", "auth_port", "35357")
     add_to_conf(nova_paste_conf, "filter:authtoken", "auth_protocol", "http")
     add_to_conf(nova_paste_conf, "filter:authtoken", "admin_tenant_name", "service")
     add_to_conf(nova_paste_conf, "filter:authtoken", "admin_user", "nova")
     add_to_conf(nova_paste_conf, "filter:authtoken", "admin_password", "nova")
-
 
     add_to_conf(nova_conf, "DEFAULT", "logdir", "/var/log/nova")
     add_to_conf(nova_conf, "DEFAULT", "lock_path", "/var/lib/nova")
@@ -374,18 +250,14 @@ def install_and_configure_neutron():
     add_to_conf(neutron_paste_conf, "filter:authtoken", "admin_user", "neutron")
     add_to_conf(neutron_paste_conf, "filter:authtoken", "admin_password", "neutron")
 	
-    add_to_conf(neutron_plugin_conf, "ml2", "type_drivers", "vlan,vxlan")
-    add_to_conf(neutron_plugin_conf, "ml2", "tenant_network_types", "vlan,vxlan")
-    add_to_conf(neutron_plugin_conf, "ml2", "mechanism_drivers", "openvswitch,linuxbridge")
-    add_to_conf(neutron_plugin_conf, "ml2_type_vlan", "network_vlan_ranges", "physnet1:2000:2999")
-    add_to_conf(neutron_plugin_conf, "ml2_type_vxlan", "vni_ranges", "500:999")
+    add_to_conf(neutron_plugin_conf, "ml2", "type_drivers", "gre")
+    add_to_conf(neutron_plugin_conf, "ml2", "tenant_network_types", "gre")
+    add_to_conf(neutron_plugin_conf, "ml2", "mechanism_drivers", "openvswitch")
+    add_to_conf(neutron_plugin_conf, "ml2_type_gre", "tunnel_id_ranges", "1:1000")
     add_to_conf(neutron_plugin_conf, "securitygroup", "firewall_driver", "neutron.agent.linux.iptables_firewall.OVSHybridIptablesFirewallDriver")
 	
-
     execute("service neutron-server restart", True)
    
-
-
 def install_and_configure_dashboard():
     execute("apt-get install openstack-dashboard -y", True)
     execute("service apache2 restart", True)
